# openapi.json
{
  "components": {
    "schemas": {
...
      "BindingCheckIn": {
        "properties": {
          "install_binding": {
            "title": "Install Binding",
            "type": "string"
          }
        },
        "required": [
          "install_binding"
        ],
        "title": "BindingCheckIn",
        "type": "object"
      },
      "BindingCheckOut": {
        "properties": {
          "bound": {
            "title": "Bound",
            "type": "boolean"
          },
          "family_mode": {
            "default": false,
            "title": "Family Mode",
            "type": "boolean"
          },
          "subject_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Name"
          }
        },
        "required": [
          "bound"
        ],
        "title": "BindingCheckOut",
        "type": "object"
      },
...
      "ChildIn": {
        "properties": {
          "academic_level": {
            "default": "smp",
            "enum": [
              "sd",
              "smp",
              "sma",
              "kuliah",
              "profesional"
            ],
            "title": "Academic Level",
            "type": "string"
          },
          "display_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "question_language": {
            "default": "id",
            "enum": [
              "id",
              "en"
            ],
            "title": "Question Language",
            "type": "string"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "ChildIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HeartbeatIn": {
        "properties": {
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Events",
            "type": "array"
          },
          "guardian_status": {
            "enum": [
              "ok",
              "degraded",
              "disabled",
              "unknown"
            ],
            "title": "Guardian Status",
            "type": "string"
          }
        },
        "required": [
          "guardian_status"
        ],
        "title": "HeartbeatIn",
        "type": "object"
      },
...
      "PairIn": {
        "properties": {
          "code": {
            "maxLength": 6,
            "minLength": 6,
            "title": "Code",
            "type": "string"
          },
          "install_binding": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Install Binding",
            "type": "string"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "platform": {
            "default": "android",
            "enum": [
              "android",
              "ios"
            ],
            "title": "Platform",
            "type": "string"
          }
        },
        "required": [
          "code",
          "install_binding"
        ],
        "title": "PairIn",
        "type": "object"
      },
      "PairingCodeOut": {
        "properties": {
          "attempts_allowed": {
            "title": "Attempts Allowed",
            "type": "integer"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "subject_id": {
            "format": "uuid",
            "title": "Subject Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "subject_id",
          "expires_at",
          "attempts_allowed"
        ],
        "title": "PairingCodeOut",
        "type": "object"
      },
...
      "SelfSubjectIn": {
        "properties": {
          "academic_level": {
            "default": "profesional",
            "enum": [
              "sd",
              "smp",
              "sma",
              "kuliah",
              "profesional"
            ],
            "title": "Academic Level",
            "type": "string"
          },
          "question_language": {
            "default": "id",
            "enum": [
              "id",
              "en"
            ],
            "title": "Question Language",
            "type": "string"
          }
        },
        "title": "SelfSubjectIn",
        "type": "object"
      },
...
      "SubjectOut": {
        "properties": {
          "academic_level": {
            "title": "Academic Level",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "academic_level",
          "kind"
        ],
        "title": "SubjectOut",
        "type": "object"
      },
      "TokenOut": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "device_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Secret"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "family_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Id"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "subject_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          },
          "user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "role"
        ],
        "title": "TokenOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Kartu waktu bermain yang dicetak dari materi belajar",
    "title": "BrainXP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/devices/check-binding": {
      "post": {
        "operationId": "check_binding_devices_check_binding_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindingCheckIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BindingCheckOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Binding",
        "tags": [
          "family"
        ]
      }
    },
    "/devices/heartbeat": {
      "post": {
        "operationId": "heartbeat_devices_heartbeat_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Heartbeat",
        "tags": [
          "family"
        ]
      }
    },
    "/devices/pair": {
      "post": {
        "operationId": "pair_device_devices_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pair Device",
        "tags": [
          "family"
        ]
      }
    },
...
    "/subjects": {
      "get": {
        "operationId": "list_subjects_subjects_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubjectOut"
                  },
                  "title": "Response List Subjects Subjects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Subjects",
        "tags": [
          "family"
        ]
      },
      "post": {
        "operationId": "create_child_subjects_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Child",
        "tags": [
          "family"
        ]
      }
    },
    "/subjects/self": {
      "post": {
        "operationId": "create_own_subject_subjects_self_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelfSubjectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Own Subject",
        "tags": [
          "family"
        ]
      }
    },
...
    "/subjects/{subject_id}/pairing-code": {
      "post": {
        "operationId": "issue_pairing_code_subjects__subject_id__pairing_code_post",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingCodeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Pairing Code",
        "tags": [
          "family"
        ]
      }
    },
...
  }
}